+++ dataprofiler/tests/profilers/profiler_options/test_unstructured_options.py
@@ -3,12 +3,15 @@ from unittest import mock
 
 from dataprofiler.profilers.json_encoder import ProfileEncoder
 from dataprofiler.profilers.profiler_options import BooleanOption, UnstructuredOptions
+from dataprofiler.tests.profilers.profiler_options.abstract_test_options import (
+    JSONDecodeTestMixin,
+)
 from dataprofiler.tests.profilers.profiler_options.test_base_option import (
     TestBaseOption,
 )
 
 
-class TestUnstructuredOptions(TestBaseOption):
+class TestUnstructuredOptions(TestBaseOption, JSONDecodeTestMixin):
 
     option_class = UnstructuredOptions
     keys = ["text", "data_labeler"]
